+++ src/primitive_list.cc
@@ -17,11 +17,11 @@
  *  LICENSE file included in the repository.
  */
 
-#include "src/PrimitiveList.h"
+#include "src/primitive_list.h"
 
 #include <QDebug>
 
-#include "src/Primitive.h"
+#include "src/primitive.h"
 
 PrimitiveList::PrimitiveList(QObject* parent)
     : QAbstractListModel{parent}, mData{} {};
