+++ test/test_beatdetect/CMakeLists.txt
@@ -2,21 +2,21 @@ project(test-beatdetect)
 
 find_package(Qt5 COMPONENTS Widgets REQUIRED)
 
-set(SOURCES ${CMAKE_SOURCE_DIR}/src/AudioFile.cpp
-            ${CMAKE_SOURCE_DIR}/src/BeatDetector.cpp)
+set(SOURCES ${CMAKE_SOURCE_DIR}/src/audio_file.cc
+            ${CMAKE_SOURCE_DIR}/src/beat_detector.cc)
 
 set(INCLUDE_DIRS  ${CMAKE_SOURCE_DIR})
 
 include_directories(${INCLUDE_DIRS})
 
 set(HEADERS ${CMAKE_SOURCE_DIR}/lib/kissfft/kissfft.hh
-            ${CMAKE_SOURCE_DIR}/src/AudioFile.h
-            ${CMAKE_SOURCE_DIR}/src/BeatDetector.h
-            ${CMAKE_SOURCE_DIR}/test/TestFolderPath.h)
+            ${CMAKE_SOURCE_DIR}/src/audio_file.h
+            ${CMAKE_SOURCE_DIR}/src/beat_detector.h
+            ${CMAKE_SOURCE_DIR}/test/test_folder_path.h)
 
 source_group("Header Files" FILES ${HEADERS})
 
-add_executable(${PROJECT_NAME} main.cpp ${SOURCES} ${HEADERS})
+add_executable(${PROJECT_NAME} main.cc ${SOURCES} ${HEADERS})
 
 target_link_libraries(  ${PROJECT_NAME}
                         gtest
